made more edits to tunnel level
added particle effects when the spider spawns, tried to fix the scale of the decorative river part (still isn't small enough) and added fire and water particle effects to the cave.
At some point the fire must dissappear and the water must appear when the power is turned back on.